backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis & Celery
redis==5.0.1
celery[redis]==5.3.4

# Vector DB
qdrant-client==1.7.0
sentence-transformers==2.2.2

# Auth & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
# hmac is built-in to Python

# LLM
google-generativeai==0.3.0
langchain==0.0.350
tiktoken==0.5.2

# JSON Schema
jsonschema==4.20.0

# Graph & ML
torch==2.1.1
torch-geometric==2.4.0
numpy==1.24.3
scikit-learn==1.3.2
pandas==2.1.4

# Payments
stripe==7.8.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Utils
python-dateutil==2.8.2
pytz==2023.3
